feat: ORMに `GORM` を使うように
- DTOに `gorm.Model` を指定することで、`ID`, `CreatedAt`, `UpdatedAt`, `DeletedAt` は継承される

+++ gin-app/domains/products/product_dao.go
@@ -7,7 +7,7 @@ import (
 )
 
 var (
-	productsDB = make(map[uint64]*Product)
+	productsDB = make(map[uint]*Product)
 )
 
 func (p *Product) Get() *errors.ApiErr {
